test: Invoke mocha via node bin script
On Windows npm generates a wrapper named mocha.cmd which is added to
%PATH%.  Unfortunately, CreateProcess() only finds .exe, .com, and .bat.
The shell would use %PATHEXT% which typically includes .cmd, but that
adds its own set of performance and quoting problems.

Instead, invoke node and pass the mocha bin script as the first
argument, which should work on all systems.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>