declare globals for eslint-plugin-n>=16.5.0
https://github.com/eslint-community/eslint-plugin-n/pull/154 removed
many global variables from eslint-plugin-n in favor of enabling the node
and es2021 environments.  Since we disable the node environment (since
our code may not be running in node) we need to declare those variables
for linting.  Do so as necessary.

Bump the required version of eslint-plugin-n to ^16.5.0 to avoid
warnings about redeclaring globals when using previous versions.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>