ci: Test earliest supported Node.js version
If we claim to support Node.js 18.0 in package.json#engines.node, we
should test that version in CI.  It's easy to inadvertently depend on
features which were backported to Node.js 18, but are not present in
early 18.x versions, thereby breaking deployments which use an older
version.  Any such breakage should be a semver major change reflected in
package.json#engines.node.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>